QR-code/requirements.txt
pillow==12.1.0
python-barcode==0.16.1
pyzbar==0.1.9
qrcode==8.2
